+++ apps/web/src/components/add-location-modal/organisms/SearchResultMap.tsx
@@ -1,0 +1,7 @@
+import * as styles from "../style.css";
+
+const SearchResultMap = () => {
+  return <div className={styles.mapContainer}></div>;
+};
+
+export default SearchResultMap;
